- Set up Rendering Context classes for OpenGL and GLFW

+++ source/Solas/Client/UserSetup.cpp
@@ -27,7 +27,7 @@ namespace Solas::Client
 
 
     /** Update is not a pure virtual function so we can throw an error if the client does not override the Update function */
-    void ClientLoop::Update(int deltaTime)
+    void ClientLoop::Update(float deltaTime)
     {
         AY_ERROR("[Ayla] Client.cpp: Update function was not defined or overridden!");
     }
